Send content_read signal from mark_read methods
Send content_read signal from mark_read methods, not the other way around

Load receivers
