+++ zds/mp/api/views.py
@@ -15,8 +15,8 @@ from zds.mp.api.permissions import IsParticipant, IsParticipantFromPrivatePost, 
     IsAloneInPrivatePost, IsAuthor
 from zds.mp.api.serializers import PrivateTopicSerializer, PrivateTopicUpdateSerializer, PrivateTopicCreateSerializer, \
     PrivatePostSerializer, PrivatePostUpdateSerializer, PrivatePostCreateSerializer
-from zds.mp.commons import LeavePrivateTopic, MarkPrivateTopicAsRead
-from zds.mp.models import PrivateTopic, PrivatePost
+from zds.mp.commons import LeavePrivateTopic
+from zds.mp.models import PrivateTopic, PrivatePost, mark_read
 from zds.utils.templatetags.interventions import interventions_privatetopics
 
 
@@ -273,7 +273,7 @@ class PrivateTopicDetailAPI(LeavePrivateTopic, RetrieveUpdateDestroyAPIView):
             return PrivateTopicUpdateSerializer
 
 
-class PrivatePostListAPI(MarkPrivateTopicAsRead, ListCreateAPIView):
+class PrivatePostListAPI(ListCreateAPIView):
     """
     Private post resource to list of a member.
     """
@@ -319,7 +319,8 @@ class PrivatePostListAPI(MarkPrivateTopicAsRead, ListCreateAPIView):
               message: Not found
         """
         response = self.list(request, *args, **kwargs)
-        self.perform_list(get_object_or_404(PrivateTopic, pk=(self.kwargs.get('pk_ptopic'))), self.request.user)
+        topic = get_object_or_404(PrivateTopic, pk=(self.kwargs.get('pk_ptopic')))
+        mark_read(topic, self.request.user)
         return response
 
     def post(self, request, *args, **kwargs):
