+++ zds/tutorialv2/utils.py
@@ -19,13 +19,14 @@ from django.core.exceptions import ObjectDoesNotExist
 from django.template.loader import render_to_string
 from django.utils.translation import ugettext_lazy as _
 from django.utils import translation
+from uuslug import slugify
 
+from zds.notification import signals
 from zds.search.models import SearchIndexContent
 from zds.tutorialv2 import REPLACE_IMAGE_PATTERN, VALID_SLUG
 from zds import settings
 from zds.settings import ZDS_APP
 from zds.utils import get_current_user
-from uuslug import slugify
 from zds.utils import slugify as old_slugify
 from zds.utils.models import Licence
 from zds.utils.templatetags.emarkdown import emarkdown
@@ -174,6 +175,7 @@ def mark_read(content, user=None):
                 content=content,
                 user=user)
             a.save()
+            signals.content_read.send(sender=content.__class__, instance=content, user=user)
 
 
 class TooDeepContainerError(ValueError):
